+++ service/src/main/java/bio/terra/workspace/app/configuration/StoragetransferConfiguration.java
@@ -1,5 +1,6 @@
-package bio.terra.workspace.service.resource.controlled.flight.clone.bucket;
+package bio.terra.workspace.app.configuration;
 
+import bio.terra.workspace.service.resource.controlled.flight.clone.bucket.StorageTransferServiceUtils;
 import com.google.api.client.googleapis.util.Utils;
 import com.google.api.services.storagetransfer.v1.Storagetransfer;
 import com.google.api.services.storagetransfer.v1.StoragetransferScopes;
@@ -15,33 +16,6 @@ import org.springframework.context.annotation.Profile;
 
 @Configuration
 public class StoragetransferConfiguration {
-  private static final Logger logger = LoggerFactory.getLogger(StoragetransferConfiguration.class);
-
-  @Bean
-  @Profile("!unit-test")
-  public GoogleCredentials getGoogleCredentials() {
-    try {
-      GoogleCredentials googleCredentials = GoogleCredentials.getApplicationDefault();
-      if (googleCredentials.createScopedRequired()) {
-        googleCredentials = googleCredentials.createScoped(StoragetransferScopes.all());
-      }
-      return googleCredentials;
-
-    } catch (IOException e) {
-      throw new BeanCreationException(e.getMessage());
-    }
-  }
-
-  /**
-   * We don't need ADC in unit tests, but we do need to provide something that looks like them in
-   * unit tests.
-   */
-  @Bean
-  @Profile("unit-test")
-  public GoogleCredentials getFakeGoogleCredentials() {
-    return GoogleCredentials.newBuilder().build();
-  }
-
   @Bean
   public Storagetransfer getStoragetransfer(GoogleCredentials googleCredentials) {
     return new Storagetransfer.Builder(
