+++ service/src/main/java/bio/terra/workspace/service/resource/controlled/flight/clone/bucket/DeleteStorageTransferServiceJobStep.java
@@ -19,12 +19,10 @@ import com.google.api.services.storagetransfer.v1.Storagetransfer;
 public class DeleteStorageTransferServiceJobStep implements Step {
 
   private final Storagetransfer storagetransfer;
-  private final CloningInstructions cloningInstructions;
 
   public DeleteStorageTransferServiceJobStep(
-      Storagetransfer storagetransfer, CloningInstructions cloningInstructions) {
+      Storagetransfer storagetransfer) {
     this.storagetransfer = storagetransfer;
-    this.cloningInstructions = cloningInstructions;
   }
 
   @Override
